shuffle some code around, move middleware code so it executes
